requirements.txt
django-recaptcha
django-avatar
